+++ CHANGELOG.md
@@ -14,6 +14,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Security: in case of vulnerabilities
 
 ## [0.1.0] - Unreleased
+
+## [0.0.3] - (https://github.com/SergeyGadzhilov/SKVM/releases/tag/0.0.3) - 2024-01-29
 ### Added
 - Logs to the sidebar [#63](https://github.com/SergeyGadzhilov/SKVM/issues/63)
 - Setting to the sidebar [#62](https://github.com/SergeyGadzhilov/SKVM/issues/62)